Fix the resolution of the cargo crates manifiest file during auto-publish
This change provides the "--manifest-path" flag to the Cargo CLI when
building packages during the Crates Auto Publish workflow. Previously,
the path would default to the current directory. This fails as we
nest different cargo packages within this repo. Explicitly providing
the path allows the build to succeed.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/365

Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>