gateway_v2: Auto publish NPM packages sequentially without workspaces
Before this change, packages were built as part of a connected workspace.
This meant that dependencies between packages were hoisted and shared.
While this is convenient for development, it can lead to unintended
consequences. Yarn will first reference local packages before it
reaches out to NPM. As a result, when building a package to publish, one
will build a dependency that represents the local layout and not that if
one had built the packages straight off NPM.

This change removes the workspace configuration before publishing to
avoid the issue, and further publishes packages in a logical and
sequential order.

For more info, please read the section on Limitations & Caveats:

https://classic.yarnpkg.com/lang/en/docs/workspaces/

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/374

Reviewed-by: Martin Riedel <web@riedel-it.de>
Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>