solana/gateway_v2: Validate fees on network creation
Gateway: Set CI Solana version to 1.14.16

This commit sets the CI build version of Solana to 1.14.16
Previously, this version was set to track "stable". This is problematic
as Solana versions are not necessarily backward compatible and may
cause our CI to break as new releases come out. Instead, a known and
working version of Solana is used.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>

Gateway: Upgrade to @coral-xyz/anchor version 0.27

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>

Gateway: Introduce InitSpace macros for sizing calculations

This change fixes an error in the size calculation of the
CreateGatekeeperAccount. Originally, the arguments for the calculation
were reversed in addition to a mistake in the underlying tallying.

To fix this, the calculation was corrected by using the InitSpace macro.
As part of this change, Anchor was upgraded to 0.27 to make use
of the new functionality.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>

Gateway: Update IDL version to 2.0.5

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>

Gateway: Provide utility method to generate a funded key

Before this change, there was a fair amount of repetition in the unit
tests for generating a funded key. This change introduces a new utility
method to help assist and reduce boilerplate code.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>