Gateway: Introduce InitSpace macros for sizing calculations
This change fixes an error in the size calculation of the
CreateGatekeeperAccount. Originally, the arguments for the calculation
were reversed in addition to a mistake in the underlying tallying.

To fix this, the calculation was corrected by using the InitSpace macro.
As part of this change, Anchor was upgraded to 0.27 to make use
of the new functionality.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>