Gateway: Set CI Solana version to 1.14.16
This commit sets the CI build version of Solana to 1.14.16
Previously, this version was set to track "stable". This is problematic
as Solana versions are not necessarily backward compatible and may
cause our CI to break as new releases come out. Instead, a known and
working version of Solana is used.

PR: https://github.com/identity-com/on-chain-identity-gateway/pull/397

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>